requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
apscheduler
httpx
pydantic-settings
python-dotenv
